+++ demo-root/demo-api/src/main/java/org/sklsft/demo/api/interfaces/reference/localization/base/CityBaseService.java
@@ -26,7 +26,7 @@ public static final String GET_LIST_URL = "/city/list";
 /**
  * load object list from region
  */
-List<CityBasicView> loadListFromRegion (Long regionId);
+List<CityBasicView> loadListFromRegion (Integer regionId);
 public static final String GET_CITY_LIST_fROM_REGION_URL = "/region/{regionId}/city/list";
 
 /**
@@ -38,7 +38,7 @@ public static final String SCROLL_URL = "/city/scroll";
 /**
  * scroll object list from region
  */
-ScrollView<CityBasicView> scrollFromRegion (Long regionId, ScrollForm<CityFilter, CitySorting> form);
+ScrollView<CityBasicView> scrollFromRegion (Integer regionId, ScrollForm<CityFilter, CitySorting> form);
 public static final String SCROLL_CITY_fROM_REGION_URL = "/region/{regionId}/city/scroll";
 
 /**
@@ -68,7 +68,7 @@ public static final String SAVE_URL = "/city";
 /**
  * save object from parent Region
  */
-Long saveFromRegion(Long regionId, CityForm cityForm);
+Long saveFromRegion(Integer regionId, CityForm cityForm);
 public static final String SAVE_FROM_REGION_URL = "/region/{regionId}/city";
 
 /**
