+++ demo-root/demo-business-component/src/main/java/org/sklsft/demo/bc/mapper/reference/time/views/full/base/CalendarDayOffFullViewBaseMapper.java
@@ -15,7 +15,7 @@ import org.sklsft.demo.model.reference.time.CalendarDayOff;
  * <br/>processed by skeleton-generator
  */
 
-public class CalendarDayOffFullViewBaseMapper extends FullViewMapper<CalendarDayOffFullView, Long, CalendarDayOffForm, CalendarDayOff> {
+public class CalendarDayOffFullViewBaseMapper extends FullViewMapper<CalendarDayOffFullView, Integer, CalendarDayOffForm, CalendarDayOff> {
 
 @Inject
 protected CalendarRightsManager calendarRightsManager;
