+++ demo-root/demo-business-component/src/main/java/org/sklsft/demo/bc/processor/organizations/base/OrganizationBaseProcessor.java
@@ -12,7 +12,7 @@ import org.sklsft.demo.repository.dao.interfaces.organizations.OrganizationDao;
  * <br/>no modification should be done to this file
  * <br/>processed by skeleton-generator
  */
-public class OrganizationBaseProcessor implements Processor<Organization, Long> {
+public class OrganizationBaseProcessor implements Processor<Organization, Integer> {
 
 /*
  * properties injected by spring
@@ -23,7 +23,7 @@ protected OrganizationDao organizationDao;
 /**
  * process save
  */
-public Long save(Organization organization) {
+public Integer save(Organization organization) {
 return organizationDao.save(organization);
 }
