+++ demo-root/demo-business-component/src/main/java/org/sklsft/demo/bc/statemanager/reference/localization/base/RegionBaseStateManager.java
@@ -9,7 +9,7 @@ import org.sklsft.demo.model.reference.localization.Region;
  * <br/>no modification should be done to this file
  * <br/>processed by skeleton-generator
  */
-public class RegionBaseStateManager implements StateManager<Region, Long> {
+public class RegionBaseStateManager implements StateManager<Region, Integer> {
 
 /**
  * can save
